Build xspice
xf86-video-qxl is provided by alpine linux. However it is not built with
the --enable-xspice flag.

I've also disabled kms as we were getting symbol lookup errors with it
enabled when running xspice. Given we are running this in a docker
container, perhaps we don't actually need kms. This needs to be
investigated more.